Initial commit of S3 events
This commit adds support for triggering a lambda function based
on an S3 event happening such as an object being created or deleted.
This is configured in chalice through a new `on_s3_event` decorator.
Chalice assumes your s3 bucket already exists.  However, if will
configure bucket configuration and function policies for you.
Given this is an existing bucket, we account for existing
notification configurations already in place, and intelligently
merge in the chalice specific lambda configuration as needed.
Same goes for `chalice delete`, it will only remove the lambda
configuration snippet it added to the S3 bucket notification config.

Before this can be merged, we need a plan for this feature
and `chalice package`.

Because you provide chalice with an existing bucket, we aren't
able to describe this in a CFN template (that is, CFN can't
adopt existing resources into a stack).  As a result, the
`chalice package` command fails.  There's a few options here.
We can implement initial support for managed resources so we
can create the bucket for you, or we can create a custom
cfn resource that can apply this config to an existing bucket.
Implementing managed resources doesn't solve the problem though
of what to do with existing buckets.  I think this is an important
enough use case that I don't want to remove support for it just
because CFN can't support this, so for me it boils down to
trying to make this work with custom resources or just not
support S3 events in `chalice package` until we add support for
managed resources.  Then at least customers would have a path
forward if they wanted to use CFN to deploy their app.  They'd
just have to accept the tradeoff that a new bucket would be
created for them.